refact(core/rulemanager): add Opt
