+++ internal/core/rulemanager/error.go
@@ -1,0 +1,9 @@
+package rulemanager
+
+import "errors"
+
+var (
+	ErrTableMissing           = errors.New("Table is missing")
+	ErrConfigMissing          = errors.New("Config is missing")
+	ErrCgroupEventChanMissing = errors.New("CgroupEvent chan is missing")
+)
