+++ internal/core/rulemanager/rulemanager.go
@@ -2,24 +2,20 @@ package rulemanager
 
 import (
 	"fmt"
-	"net"
 	"regexp"
 
 	"github.com/black-desk/deepin-network-proxy-manager/internal/config"
 	"github.com/black-desk/deepin-network-proxy-manager/internal/core/rulemanager/table"
-	"github.com/black-desk/deepin-network-proxy-manager/internal/inject"
-	. "github.com/black-desk/deepin-network-proxy-manager/internal/log"
 	"github.com/black-desk/deepin-network-proxy-manager/internal/types"
 	"github.com/black-desk/deepin-network-proxy-manager/pkg/location"
 	"github.com/vishvananda/netlink"
-	"golang.org/x/sys/unix"
 )
 
 type RuleManager struct {
-	EventChan <-chan *types.CgroupEvent `inject:"true"`
+	cgroupEventChan <-chan *types.CgroupEvent `inject:"true"`
 
-	Nft *table.Table   `inject:"true"`
-	Cfg *config.Config `inject:"true"`
+	nft *table.Table   `inject:"true"`
+	cfg *config.Config `inject:"true"`
 
 	matchers []*regexp.Regexp
 
@@ -27,7 +23,7 @@ type RuleManager struct {
 	route *netlink.Route
 }
 
-func New(container *inject.Container) (m *RuleManager, err error) {
+func New(opts ...Opt) (ret *RuleManager, err error) {
 	defer func() {
 		if err == nil {
 			return
@@ -39,14 +35,16 @@ func New(container *inject.Container) (m *RuleManager, err error) {
 		)
 	}()
 
-	m = &RuleManager{}
-	err = container.Fill(m)
-	if err != nil {
-		return
+	m := &RuleManager{}
+	for i := range opts {
+		m, err = opts[i](m)
+		if err != nil {
+			return
+		}
 	}
 
-	for i := range m.Cfg.Rules {
-		regex := m.Cfg.Rules[i].Match
+	for i := range m.cfg.Rules {
+		regex := m.cfg.Rules[i].Match
 		var matcher *regexp.Regexp
 		matcher, err = regexp.Compile(regex)
 		if err != nil {
@@ -56,197 +54,47 @@ func New(container *inject.Container) (m *RuleManager, err error) {
 		m.matchers = append(m.matchers, matcher)
 	}
 
+	ret = m
 	return
 }
 
-func (m *RuleManager) Run() (err error) {
-	defer func() {
-		if err == nil {
+type Opt func(m *RuleManager) (ret *RuleManager, err error)
+
+func WithTable(t *table.Table) Opt {
+	return func(m *RuleManager) (ret *RuleManager, err error) {
+		if t == nil {
+			err = ErrTableMissing
 			return
 		}
 
-		err = fmt.Errorf(location.Capture()+
-			"Error occurs while running the nftable rules manager:\n%w",
-			err,
-		)
-	}()
-
-	defer m.removeNftableRules()
-	err = m.initializeNftableRuels()
-	if err != nil {
+		m.nft = t
+		ret = m
 		return
 	}
-
-	defer m.removeRoute()
-	err = m.addRoute()
-	if err != nil {
-		return
-	}
-
-	defer m.removeRule()
-	err = m.addRule()
-	if err != nil {
-		return
-	}
-
-	for event := range m.EventChan {
-		switch event.EventType {
-		case types.CgroupEventTypeNew:
-			m.handleNewCgroup(event.Path)
-		case types.CgroupEventTypeDelete:
-			m.handleDeleteCgroup(event.Path)
-		}
-	}
-	return
 }
 
-func (m *RuleManager) initializeNftableRuels() (err error) {
-	defer func() {
-		if err == nil {
+func WithConfig(c *config.Config) Opt {
+	return func(m *RuleManager) (ret *RuleManager, err error) {
+		if c == nil {
+			err = ErrConfigMissing
 			return
 		}
 
-		err = fmt.Errorf(location.Capture()+
-			"Failed to initialize nftable ruels:\n%w",
-			err,
-		)
-	}()
-
-	for _, tp := range m.Cfg.TProxies {
-		// NOTE(black_desk): Same as `addChainForProxy`.
-		_ = m.Nft.AddChainAndRulesForTProxy(tp)
+		m.cfg = c
+		ret = m
+		return
 	}
-
-	err = m.Nft.FlushInitialContent()
-	return
 }
 
-func (m *RuleManager) removeNftableRules() (err error) {
-	err = m.Nft.Clear()
-	return
-}
-
-func (m *RuleManager) addRoute() (err error) {
-	defer func() {
-		if err == nil {
+func WithCgroupEventChan(ch <-chan *types.CgroupEvent) Opt {
+	return func(m *RuleManager) (ret *RuleManager, err error) {
+		if ch == nil {
+			err = ErrCgroupEventChanMissing
 			return
 		}
 
-		err = fmt.Errorf(location.Capture()+
-			"Failed to add route:\n%w",
-			err,
-		)
-	}()
-
-	var iface *net.Interface
-	iface, err = net.InterfaceByName("lo")
-	if err != nil {
+		m.cgroupEventChan = ch
+		ret = m
 		return
-	}
-
-	route := &netlink.Route{
-		LinkIndex: iface.Index,
-		Scope:     unix.RT_SCOPE_HOST,
-		Dst:       &net.IPNet{IP: net.IPv4zero, Mask: net.CIDRMask(0, 32)},
-		Protocol:  unix.RTPROT_BOOT,
-		Table:     m.Cfg.RouteTable,
-		Type:      unix.RTN_LOCAL,
-	}
-
-	err = netlink.RouteAdd(route)
-	if err != nil {
-		return
-	}
-
-	m.route = route
-
-	return
-}
-
-func (m *RuleManager) removeRoute() {
-	err := netlink.RouteDel(m.route)
-
-	if err == nil {
-		return
-	}
-
-	Log.Warnw("Failed to delete route", "error", err)
-
-	return
-}
-
-func (m *RuleManager) addRule() (err error) {
-	defer func() {
-		if err == nil {
-			return
-		}
-
-		err = fmt.Errorf(location.Capture()+
-			"Failed to add route rule:\n%w",
-			err,
-		)
-	}()
-
-	rule := netlink.NewRule()
-	rule.Family = netlink.FAMILY_ALL
-	rule.Mark = int(m.Cfg.Mark) // WARN(black_desk): ???
-	rule.Table = m.Cfg.RouteTable
-
-	err = netlink.RuleAdd(rule)
-	if err != nil {
-		return
-	}
-
-	m.rule = rule
-
-	return
-}
-
-func (m *RuleManager) removeRule() {
-	err := netlink.RuleDel(m.rule)
-
-	if err == nil {
-		return
-	}
-
-	Log.Warnw("Failed to delete rule", "error", err)
-
-	return
-}
-
-func (m *RuleManager) handleNewCgroup(path string) {
-	var target table.Target
-	for i := range m.matchers {
-		if !m.matchers[i].Match([]byte(path)) {
-			continue
-		}
-
-		if m.Cfg.Rules[i].Direct {
-			target.Op = table.TargetDirect
-		} else if m.Cfg.Rules[i].Drop {
-			target.Op = table.TargetDrop
-		} else if m.Cfg.Rules[i].Proxy != "" {
-			target.Op = table.TargetTProxy
-			target.Chain = m.Cfg.Proxies[m.Cfg.Rules[i].Proxy].TProxy.Name
-		} else if m.Cfg.Rules[i].TProxy != "" {
-			target.Op = table.TargetTProxy
-			target.Chain = m.Cfg.TProxies[m.Cfg.Rules[i].TProxy].Name
-		} else {
-			panic("this should never happened.")
-		}
-
-		break
-	}
-
-	err := m.Nft.AddCgroup(path, &target)
-	if err != nil {
-		Log.Errorw("Failed to update nft for new cgroup", "error", err)
-	}
-}
-
-func (m *RuleManager) handleDeleteCgroup(path string) {
-	err := m.Nft.RemoveCgroup(path)
-	if err != nil {
-		Log.Errorw("Failed to update nft for removed cgroup", "error", err)
 	}
 }
